fix(runtime): rename contains_async to contains_await
